requirements/requirements_client.txt
websockets>=10.0
opencv-python>=4.0.0.21
numpy>=1.19.0
requests>=2.25.0
Pillow>=8.0.0
tkinterdnd2~=0.4.2

